@@ -4,8 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [tool.black]
 line-length = 119
-target-version = ["py36"]
-include = 'djangocms_page_sitemap/*py'
+target-version = ["py310"]
 
 [tool.towncrier]
 package = "djangocms_page_sitemap"
@@ -28,3 +27,22 @@ verbose = 0
 quiet = false
 whitelist-regex = []
 color = true
+
+[tool.isort]
+profile = "black"
+combine_as_imports = true
+default_section = "THIRDPARTY"
+force_grid_wrap = 0
+include_trailing_comma = true
+known_first_party = "djangocms_page_sitemap"
+line_length = 119
+multi_line_output = 3
+use_parentheses = true
+
+[tool.ruff]
+ignore = []
+line-length = 119
+target-version = "py310"
+
+[tool.ruff.mccabe]
+max-complexity = 10
