@@ -1,10 +1,9 @@
 from cms.utils.conf import get_cms_setting
 from django.conf import settings
-from django.conf.urls import include
 from django.conf.urls.i18n import i18n_patterns
 from django.contrib import admin
 from django.contrib.staticfiles.urls import staticfiles_urlpatterns
-from django.urls import re_path
+from django.urls import include, path, re_path
 from django.views.static import serve
 
 from djangocms_page_sitemap import sitemap_urls
@@ -22,12 +21,12 @@ urlpatterns = [
         serve,
         {"document_root": get_cms_setting("MEDIA_ROOT"), "show_indexes": True},
     ),
-    re_path(r"^", include(sitemap_urls)),
+    path("", include(sitemap_urls)),
 ]
 
 urlpatterns += staticfiles_urlpatterns()
 
 urlpatterns += i18n_patterns(
     re_path(r"^admin/", admin.site.urls),
-    re_path(r"^", include("cms.urls")),
+    path("", include("cms.urls")),
 )
